+++ src/helpers/commandHelper.ts
@@ -1,0 +1,28 @@
+import { OCAPIService } from '../services/OCAPIService';
+import SitePreferencesHelper from './SitePreferencesHelper';
+import { MetadataNode } from '../components/MetadataNode';
+
+/**
+ * @class
+ * @classdesc - A helper class for handling command actions that are separate
+ *    from the view provider functionality.
+ */
+export default class CommandHelper {
+  private service: OCAPIService;
+  private prefsHelper: SitePreferencesHelper;
+
+  constructor() {
+    this.service = new OCAPIService();
+    this.prefsHelper = new SitePreferencesHelper(this.service);
+  }
+
+  /**
+   * Handles the command action to set the value of a SitePreferences custom
+   * attribute to the value specified by the user.
+   * @param {MetadataNode} element - The selected element when the command was
+   *    invoked to set the pref value.
+   */
+  public async setPref(element: MetadataNode) {
+    /** @todo:  */
+  }
+}