+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to the "sfcc-metadata-explorer" extension will be documented in this file.
 
-## Unreleased
+## Unreleased Changes
+- Added listener to extension.ts to handle the context menu action.
+- Added command `extension.sfccexplorer.sitepreference.setvalue` to the package.json for setting preference values.
+- Added API config for the `PATCH` call to OCAPI `Sites` resource to set preference values.
+- Added feature to readme.md features list.
+- Updated CHANGELOG.md with notes of changes..
+
+## 0.8.1 (3/21/20)
+### Patch Release
 - Updated version of npm package `webpack` to **^4.42.0**.
 - Updated version of npm package `webpack-cli` to **^3.3.11**.
 - Updated version to **0.8.1**.